requirements.txt
torch
transformers
datasets
evaluate
scikit-learn
onnx
onnxruntime
optimum
pyyaml
pandas
numpy
huggingface_hub
python-dotenv
